+++ proxy/server/conn_query.go
@@ -132,7 +132,7 @@ func (c *ClientConn) getBackendConn(n *backend.Node, fromSlave bool) (co *backen
 			c.txConns[n] = co
 		}
 	}
-	//todo, set conn charset, etc...
+
 	if err = co.UseDB(c.db); err != nil {
 		return
 	}
@@ -144,7 +144,7 @@ func (c *ClientConn) getBackendConn(n *backend.Node, fromSlave bool) (co *backen
 	return
 }
 
-/*获取shard的conn，第一个参数表示是不是select*/
+//获取shard的conn，第一个参数表示是不是select
 func (c *ClientConn) getShardConns(fromSlave bool, plan *router.Plan) (map[string]*backend.BackendConn, error) {
 	var err error
 	if plan == nil || len(plan.RouteNodeIndexs) == 0 {
@@ -301,7 +301,6 @@ func (c *ClientConn) closeShardConns(conns map[string]*backend.BackendConn, roll
 		if rollback {
 			co.Rollback()
 		}
-
 		co.Close()
 	}
 }
@@ -331,198 +330,6 @@ func (c *ClientConn) newEmptyResultset(stmt *sqlparser.Select) *mysql.Resultset 
 	r.RowDatas = make([]mysql.RowData, 0)
 
 	return r
-}
-
-func (c *ClientConn) GetTransExecNode(tokens []string, sql string) (*backend.Node, error) {
-	var execNode *backend.Node
-	var err error
-
-	tokensLen := len(tokens)
-	if 2 <= tokensLen {
-		if tokens[0][0] == mysql.COMMENT_PREFIX {
-			nodeName := strings.Trim(tokens[0], mysql.COMMENT_STRING)
-			if c.schema.nodes[nodeName] != nil {
-				execNode = c.schema.nodes[nodeName]
-			}
-		}
-	}
-
-	if execNode == nil {
-		execNode, _, err = c.GetExecNode(tokens, sql)
-		if err != nil {
-			return nil, err
-		}
-		return execNode, nil
-	}
-	if len(c.txConns) == 1 && c.txConns[execNode] == nil {
-		return nil, errors.ErrTransInMulti
-	}
-	return execNode, nil
-}
-
-func (c *ClientConn) GetExecNode(tokens []string,
-	sql string) (*backend.Node, bool, error) {
-	var execNode *backend.Node
-	var fromSlave bool
-
-	schema := c.proxy.schema
-	rules := schema.rule.Rules
-
-	tokensLen := len(tokens)
-	if 0 < tokensLen {
-		tokenId, ok := mysql.PARSE_TOKEN_MAP[strings.ToLower(tokens[0])]
-		if ok == true {
-			switch tokenId {
-			case mysql.TK_ID_SELECT, mysql.TK_ID_DELETE:
-				if len(rules) == 0 {
-					if tokenId == mysql.TK_ID_SELECT {
-						fromSlave = true
-					}
-					break
-				}
-				for i := 1; i < tokensLen; i++ {
-					if strings.ToLower(tokens[i]) == mysql.TK_STR_FROM {
-						if i+1 < tokensLen {
-							tableName := sqlparser.GetTableName(tokens[i+1])
-
-							if _, ok := rules[tableName]; ok {
-								return nil, false, nil
-							} else {
-								if tokenId == mysql.TK_ID_SELECT {
-									fromSlave = true
-								}
-							}
-						}
-					}
-				}
-			case mysql.TK_ID_INSERT, mysql.TK_ID_REPLACE:
-				if len(rules) == 0 {
-					break
-				}
-				for i := 0; i < tokensLen; i++ {
-					if strings.ToLower(tokens[i]) == mysql.TK_STR_INTO {
-						if i+1 < tokensLen {
-							tableName := sqlparser.GetInsertTableName(tokens[i+1])
-							if _, ok := rules[tableName]; ok {
-								return nil, false, nil
-							}
-						}
-					}
-				}
-			case mysql.TK_ID_UPDATE:
-				if len(rules) == 0 {
-					break
-				}
-				for i := 0; i < tokensLen; i++ {
-					if strings.ToLower(tokens[i]) == mysql.TK_STR_SET {
-						tableName := sqlparser.GetTableName(tokens[i-1])
-						if _, ok := rules[tableName]; ok {
-							return nil, false, nil
-						}
-					}
-				}
-			case mysql.TK_ID_SET:
-				if len(tokens) < 2 {
-					break
-				}
-				tmp1 := strings.Split(sql, "=")
-				tmp2 := strings.Split(tmp1[0], " ")
-				secondWord := strings.ToLower(tmp2[1])
-				if secondWord == mysql.TK_STR_NAMES ||
-					secondWord == mysql.TK_STR_RESULTS ||
-					secondWord == mysql.TK_STR_CLIENT ||
-					secondWord == mysql.TK_STR_CONNECTION ||
-					secondWord == mysql.TK_STR_AUTOCOMMIT {
-					return nil, false, nil
-				}
-			default:
-				return nil, false, nil
-			}
-		}
-	}
-	//get node
-	if 2 <= tokensLen {
-		if tokens[0][0] == mysql.COMMENT_PREFIX {
-			nodeName := strings.Trim(tokens[0], mysql.COMMENT_STRING)
-			if c.schema.nodes[nodeName] != nil {
-				execNode = c.schema.nodes[nodeName]
-			}
-			//select
-			if mysql.PARSE_TOKEN_MAP[tokens[1]] == mysql.TK_ID_SELECT {
-				fromSlave = true
-			}
-		}
-	}
-
-	if execNode == nil {
-		defaultRule := c.schema.rule.DefaultRule
-		if len(defaultRule.Nodes) == 0 {
-			return nil, false, errors.ErrNoDefaultNode
-		}
-		execNode = c.proxy.GetNode(defaultRule.Nodes[0])
-	}
-
-	return execNode, fromSlave, nil
-}
-
-//返回true表示已经处理，false表示未处理
-func (c *ClientConn) preHandleShard(sql string) (bool, error) {
-	var rs []*mysql.Result
-	var err error
-
-	var execNode *backend.Node
-	var fromSlave bool = false
-
-	if len(sql) == 0 {
-		return false, errors.ErrCmdUnsupport
-	}
-
-	tokens := strings.Fields(sql)
-	if len(tokens) == 0 {
-		return false, errors.ErrCmdUnsupport
-	}
-
-	if c.isInTransaction() {
-		execNode, err = c.GetTransExecNode(tokens, sql)
-	} else {
-		execNode, fromSlave, err = c.GetExecNode(tokens, sql)
-	}
-
-	if err != nil {
-		return false, err
-	}
-	//need shard sql
-	if execNode == nil {
-		return false, nil
-	}
-	//execute in Master DB
-	conn, err := c.getBackendConn(execNode, fromSlave)
-	defer c.closeConn(conn, false)
-	if err != nil {
-		return false, err
-	}
-	rs, err = c.executeInNode(conn, sql, nil)
-	if err != nil {
-		return false, err
-	}
-
-	if len(rs) == 0 {
-		msg := fmt.Sprintf("result is empty")
-		golog.Error("ClientConn", "handleUnsupport", msg, c.connectionId)
-		return false, mysql.NewError(mysql.ER_UNKNOWN_ERROR, msg)
-	}
-
-	if rs[0].Resultset != nil {
-		err = c.writeResultset(c.status, rs[0].Resultset)
-	} else {
-		err = c.writeOK(rs[0])
-	}
-
-	if err != nil {
-		return false, err
-	}
-
-	return true, nil
 }
 
 func (c *ClientConn) handleExec(stmt sqlparser.Statement, args []interface{}) error {
